+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ## Version 0
 ### Version 0.4.0
+* New get_user_gender method for getting the gender of the provided user.
 * New get_current_user_meta for getting userlists about the currently logged in user.
 * Fix all userlist methods to work without supplying a username.
 * New get_registration_time method to get when the user registered.